Refatoração.

+++ src/main/java/br/com/delta/Main.java
@@ -17,7 +17,7 @@ public class Main {
 			HashMap parametros = new HashMap();
 
 			if (args.length == 0) {
-				parametros.put("ID", new Integer(1));
+				parametros.put("ID", new Integer(103));
 			} else {
 				parametros.put("ID", new Integer(args[0]));
 			}
